backend/requirements.txt
flask==3.1.2
flask-cors==4.0.0
Flask-SQLAlchemy==3.0.3
psycopg2-binary==2.9.6
python-dotenv==1.0.0
gunicorn==20.1.0
requests==2.32.3
